+++ foxes/output/__init__.py
@@ -2,5 +2,5 @@ from .output import Output
 from .farm_layout import FarmLayoutOutput
 from .farm_results_eval import FarmResultsEval
 from .flow_plots_2d import FlowPlots2D
-from .rose_plot import RosePlotOutput, AmbientRosePlotOutput
+from .rose_plot import RosePlotOutput, StatesRosePlotOutput
 from .results_writer import ResultsWriter